requirements.txt
langchain
python-dotenv
openai
flask